Fix tipo identificacion

+++ src/main/java/com/coagronet/tipoIdentificacion/services/TipoIdentificacionService.java
@@ -2,51 +2,68 @@ package com.coagronet.tipoIdentificacion.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.coagronet.tipoIdentificacion.TipoIdentificacion;
+import com.coagronet.tipoIdentificacion.dtos.TipoIdentificacionDTO;
+import com.coagronet.tipoIdentificacion.mappers.TipoIdentificacionMapper;
 import com.coagronet.tipoIdentificacion.repositories.TipoIdentificacionRepository;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
+@RequiredArgsConstructor
 public class TipoIdentificacionService {
 
-    @Autowired
-    private TipoIdentificacionRepository tipoIdentificacionRepository;
+    private final TipoIdentificacionRepository tipoIdentificacionRepository;
+    private final TipoIdentificacionMapper tipoIdentificacionMapper;
 
-    // Create
-    public TipoIdentificacion add(TipoIdentificacion tipoIdentificacion) {
-        return tipoIdentificacionRepository.save(tipoIdentificacion);
+    public List<TipoIdentificacionDTO> findAll() {
+        return tipoIdentificacionRepository.findAll()
+                .stream()
+                .map(tipoIdentificacionMapper::toDTO)
+                .collect(Collectors.toList());
     }
 
-    // Read all
-    public List<TipoIdentificacion> getAll() {
-        return tipoIdentificacionRepository.findAll();
+    public List<TipoIdentificacionDTO> findAllAvailable() {
+        return tipoIdentificacionRepository.findByEstadoIdNotOrderByIdAsc(2L)
+                .stream()
+                .map(tipoIdentificacionMapper::toDTO)
+                .collect(Collectors.toList());
     }
 
-    // Read by ID
-    public TipoIdentificacion getById(Integer id) {
-        Optional<TipoIdentificacion> tipoIdentificacion = tipoIdentificacionRepository.findById(id);
-        return tipoIdentificacion.orElse(null);
+    public Optional<TipoIdentificacionDTO> findById(Long requestedId) {
+        return tipoIdentificacionRepository.findById(requestedId)
+                .map(tipoIdentificacionMapper::toDTO);
     }
 
-    // Update
-    public TipoIdentificacion update(Integer id, TipoIdentificacion tipoIdentificacionDetails) {
-        TipoIdentificacion tipoIdentificacion = getById(id);
-        if (tipoIdentificacion != null) {
-            tipoIdentificacion.setNombre(tipoIdentificacionDetails.getNombre());
-            tipoIdentificacion.setDescripcion(tipoIdentificacionDetails.getDescripcion());
-            tipoIdentificacion.setEstado(tipoIdentificacionDetails.getEstado());
+    public TipoIdentificacionDTO create(TipoIdentificacionDTO tipoIdentificacionDTO) {
+        tipoIdentificacionDTO.setId(null);
+        return tipoIdentificacionMapper.toDTO(
+                tipoIdentificacionRepository.save(
+                        tipoIdentificacionMapper.toEntity(tipoIdentificacionDTO)));
+    }
 
-            return tipoIdentificacionRepository.save(tipoIdentificacion);
+    public boolean update(Long requestedId, TipoIdentificacionDTO tipoIdentificacionDTO) {
+        if (tipoIdentificacionRepository.existsById(requestedId)) {
+            tipoIdentificacionDTO.setId(requestedId);
+            TipoIdentificacion tipoIdentificacion = tipoIdentificacionMapper.toEntity(tipoIdentificacionDTO);
+            tipoIdentificacionRepository.save(tipoIdentificacion);
+            return true;
+        } else {
+            return false;
         }
-        return null;
     }
 
-    // Delete
-    public void delete(Integer id) {
-        tipoIdentificacionRepository.deleteById(id);
+    public boolean delete(Long id) {
+        if (tipoIdentificacionRepository.existsById(id)) {
+            tipoIdentificacionRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
 }
